pyproject.toml
[tool.poetry]
name = "st-atlas-datasets"
version = "0.1.0"
description = ""
authors = ["Ouail Bendidi <ouail.bendidi@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "st_atlas_datasets", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = ">=2,<3"
pydantic-settings = "^2.1.0"
rich = "^13.7.0"
parsel = "^1.8.1"
typing-extensions = "^4.8.0"
fsspec = {extras = ["http"], version = ">=2023.1.0"}
pandas = "^2.1.3"
pillow = "^10.1.0"
tifffile = "^2023.9.26"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
ruff = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff.isort]
known-first-party = ["st_atlas_datasets"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q -vv --cov=src/st_atlas_datasets --cov-report=term"

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
omit = ["**/tests/test_*.py"]
fail_under = 90
